Factor out Git modules
